requirements.txt
# AI Employee - Python Dependencies
# Bronze Tier Requirements

# Core dependencies
watchdog>=4.0.0          # File system watching for orchestrator

# Gmail API (optional - for Gmail watcher)
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0

# Utilities
python-dotenv>=1.0.0     # Environment variable management

# Development
pytest>=7.0.0            # Testing (optional)
